requirements.txt
langchain==0.3.27
langchain-openai==0.3.33
langchain-community==0.3.27
faiss-cpu~=1.13.2
streamlit~=1.47.0
pypdf~=6.7.4
GitPython~=3.1.44
gitdb~=4.0.12
h11~=0.14.0
defusedxml~=0.7.1
ipython~=8.31.0
numpy~=2.1.3
pillow~=11.1.0
pip~=24.0
typing_extensions~=4.15.0
attrs~=24.3.0
distro~=1.9.0
protobuf~=5.29.5
filelock~=3.18.0
pytz~=2025.1
toml~=0.10.2
requests~=2.32.3
rich~=14.0.0
keras~=3.10.0
notebook~=7.3.2
PyYAML~=6.0.2
idna~=3.10
multidict~=6.7.1
propcache~=0.4.1
sniffio~=1.3.1
torch~=2.7.0
scipy~=1.15.3
packaging~=24.2
smmap~=5.0.2
click~=8.2.1
Pygments~=2.18.0
httpcore~=1.0.7
httpx~=0.28.1
zstandard~=0.25.0
jiter~=0.13.0
charset-normalizer~=3.4.1
setuptools~=75.6.0
overrides~=7.7.0
cffi~=1.17.1
regex~=2024.11.6
altair~=5.5.0
jsonschema~=4.23.0
narwhals~=1.47.1
pandas~=2.2.3
pyarrow~=20.0.0
Jinja2~=3.1.5
referencing~=0.35.1
traitlets~=5.14.3
schema~=0.7.8
MarkupSafe~=3.0.2
openai~=1.109.1
pydantic~=2.12.0
anyio~=4.7.0
pydantic_core~=2.41.1
orjson~=3.11.5
openpyxl~=3.1.5
fsspec~=2025.5.1
SQLAlchemy~=2.0.47
matplotlib~=3.10.1
python-dateutil~=2.9.0.post0
pydeck~=0.9.1
xxhash~=3.6.0
yarl~=1.22.0
aiohttp~=3.13.3
aiosignal~=1.4.0
frozenlist~=1.8.0
aiohappyeyeballs~=2.6.1
certifi~=2025.6.15
tzdata~=2025.1
tornado~=6.4.2
urllib3~=2.3.0
six~=1.17.0
duckdb~=1.3.2
annotated-types~=0.7.0
tenacity~=9.1.2
tiktoken~=0.12.0
langchain-core~=0.3.83
langsmith~=0.6.2
langchain-text-splitters~=0.3.11
psutil~=6.1.1
uuid_utils~=0.13.0
requests-toolbelt~=1.0.0
cachetools~=6.1.0
blinker~=1.9.0
sympy~=1.13.3
webcolors~=24.11.1
jsonpointer~=3.0.0
uri-template~=1.3.0
isoduration~=20.11.0
fqdn~=1.5.1
rfc3339-validator~=0.1.4
jsonschema-specifications~=2024.10.1
marshmallow~=3.26.2
jsonpatch~=1.33
transformers~=4.52.3
typing-inspect~=0.9.0
dataclasses-json~=0.6.7
pydantic-settings~=2.13.1
python-dotenv~=1.2.1
mypy_extensions~=1.1.0
httpx-sse~=0.4.3
beautifulsoup4~=4.12.3
networkx~=3.4.2